Fixed subscription, now they works!
I've discovered that WebSocket worked on a different endpoint of the
GraphQL API exposed by the server, which are:
`ws://host:port/subscriptions`.
Client now is working correctly as expected, the only problem is when
disconnecting the client that takes about a minute.
If the project stays in local host, it could be fine to add a short
timeout and try to avoid reconnections due to network errors, because
being on local host, would make no sense to retry in case of suche
errors.